Iterated on Change Password with Major Code Refactoring
