+++ src/components/Profile/ChangePassword/ChangePasswordWindow.js
@@ -1,38 +1,32 @@
-import React from 'react';
+import React, { useState } from 'react';
 import {
   Box, Button, Dialog,
 } from '@material-ui/core';
-import useStyles from './ChangePasswordStyles';
-import ChangePasswordForm from './ChangePasswordForm';
-import ChangePasswordSuccess from './ChangePasswordSuccess';
+import PasswordChangeScreen from './PasswordChangeScreen';
+import SuccessChangeScreen from './SuccessChangeScreen';
 
 export default function ChangePasswordWindow() {
-  const classes = useStyles();
-  const [open, setOpen] = React.useState(false);
-  const [isSubmitted, setSubmit] = React.useState(false);
-  const [passwords, setPasswords] = React.useState({
-    oldP: '',
-    confirmP: '',
-    newP: '',
+  const [open, setOpen] = useState(false);
+  const [isSubmitted, setSubmitted] = useState(false);
+  const [passwords, setPasswords] = useState({
+    oldPassword: '',
+    confirmPassword: '',
+    newPassword: '',
   });
 
   const handleOpen = () => {
     setOpen(true);
   };
-
   const handleClose = () => {
     setOpen(false);
   };
-
   const handlePasswordChange = (e) => {
     const { name, value } = e.target;
     setPasswords({ ...passwords, [name]: value });
   };
 
-  const validate = () => true;
-
   return (
-    <Box m={2} p={2} className={classes.box}>
+    <Box m={2} p={2}>
       <Button variant="outlined" color="primary" onClick={handleOpen}>
         Change Password
       </Button>
@@ -43,15 +37,14 @@ export default function ChangePasswordWindow() {
         fullWidth
       >
         {!isSubmitted && (
-          <ChangePasswordForm
+          <PasswordChangeScreen
             handleClose={handleClose}
             handlePasswordChange={handlePasswordChange}
             passwords={passwords}
-            validateForm={validate}
-            setSubmit={setSubmit}
+            setSubmit={setSubmitted}
           />
         )}
-        {isSubmitted && <ChangePasswordSuccess handleClose={handleClose} />}
+        {isSubmitted && <SuccessChangeScreen handleClose={handleClose} />}
       </Dialog>
     </Box>
   );
