--- src/components/Profile/ChangePasswordStyles.js
@@ -1,56 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles({
-  leftHeading: {
-    fontFamily: 'Work Sans',
-    fontSize: '28px',
-    fontWeight: '700',
-    color: '#373737',
-    textAlign: 'left',
-  },
-  errorText: {
-    fontFamily: 'Work Sans',
-    fontSize: '14px',
-    color: '#ff0000',
-    textAlign: 'center',
-    marginTop: '10px',
-    margin: '0px',
-  },
-  form: {
-    width: '400px',
-    fontFamily: 'Work Sans',
-    display: 'flex',
-    flexDirection: 'column',
-    margin: 'auto',
-  },
-  formControl: {
-    margin: '0px',
-    marginBottom: '10px',
-    width: '100%',
-  },
-  inputLabel: {
-    fontFamily: 'Work Sans',
-    fontWeight: '600',
-    color: '#373737',
-    '&:after': {
-      content: '" *"',
-      color: '#f00',
-    },
-  },
-  placeholder: {
-    fontFamily: 'Work Sans',
-  },
-  changePasswordButton: {
-    fontFamily: 'Work Sans',
-  },
-  cancelButton: {
-    fontFamily: 'Work Sans',
-    textTransform: 'none',
-    fontWeight: '600',
-  },
-  centerAlignDialogActions: {
-    justifyContent: 'center',
-  },
-});
-
-export default useStyles;
