+++ src/components/Profile/ProfileScreen.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
+import ChangePasswordWindow from './ChangePassword/ChangePasswordWindow';
 
 const useStyles = makeStyles({
   root: {
@@ -12,7 +13,9 @@ export default function ProfileScreen() {
   const classes = useStyles();
   return (
     <div className={classes.root}>
+      {/* TODO: Complete rest of profile screen */}
       <h1>Profile Screen</h1>
+      <ChangePasswordWindow />
     </div>
   );
 }
